i added all the functionalities but needs lot of touches everywhere.
all the functionalities work but needs lots of touches everywhere in order to be a professional website. The database is working perfectly. Little bit of dummy data is wrong because i put it in the starting when i created the DB. But when the user uses the website, there are no errors as far as i could notice.